# perf: Improve push.RetentionPeriodToString() efficiency