**What this PR does / why we need it**:

This PR improves the efficiency of `loghttp/push.RetentionPeriodToString()` which is used in `loghttp/push.go` and can be used when implementing the `push.UsageTracker` interface. The goal of this PR is an easy, if small, performance improvement for Distributors.

```bash
$ ~/go/bin/benchstat ~/Desktop/old_1.txt ~/Desktop/new_1.txt
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/v3/pkg/loghttp/push
cpu: Apple M3
                                                
                                                │                 sec/op                 │    sec/op      vs base                 │
RetentionPeriodToString/744h-8                                              46.03n ± ∞ ¹    12.76n ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/840h-8                                              43.89n ± ∞ ¹    12.68n ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/2232h-8                                             52.16n ± ∞ ¹    13.00n ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/8784h-8                                             46.23n ± ∞ ¹    13.05n ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/1000h-8                                             46.20n ± ∞ ¹    13.05n ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/zero_retention_period-8                            1.3720n ± ∞ ¹   0.2474n ± ∞ ¹        ~ (p=1.000 n=1) ²
geomean                                                                     26.00n          6.677n        -74.32%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                                
                                                │                  B/op                  │     B/op       vs base                 │
RetentionPeriodToString/744h-8                                              16.000 ± ∞ ¹     3.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/840h-8                                              16.000 ± ∞ ¹     3.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/2232h-8                                             16.000 ± ∞ ¹     4.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/8784h-8                                             16.000 ± ∞ ¹     4.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/1000h-8                                             16.000 ± ∞ ¹     4.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/zero_retention_period-8                              0.000 ± ∞ ¹     0.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
geomean                                                                                ⁴                  -71.38%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal
⁴ summaries must be >0 to compute geomean

                                                
                                                │               allocs/op                │   allocs/op    vs base                 │
RetentionPeriodToString/744h-8                                               2.000 ± ∞ ¹     1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/840h-8                                               2.000 ± ∞ ¹     1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/2232h-8                                              2.000 ± ∞ ¹     1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/8784h-8                                              2.000 ± ∞ ¹     1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/1000h-8                                              2.000 ± ∞ ¹     1.000 ± ∞ ¹        ~ (p=1.000 n=1) ²
RetentionPeriodToString/zero_retention_period-8                              0.000 ± ∞ ¹     0.000 ± ∞ ¹        ~ (p=1.000 n=1) ³
geomean                                                                                ⁴                  -43.88%               ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal
⁴ summaries must be >0 to compute geomean
```

**Which issue(s) this PR fixes**:

N/A

**Special notes for your reviewer**:

N/A

**Checklist**
- [x] Reviewed the [`CONTRIBUTING.md`](https://github.com/grafana/loki/blob/main/CONTRIBUTING.md) guide (**required**)
- [ ] Documentation added
- [x] Tests updated
- [x] Title matches the required conventional commits format, see [here](https://www.conventionalcommits.org/en/v1.0.0/)
  - **Note** that Promtail is considered to be feature complete, and future development for logs collection will be in [Grafana Alloy](https://github.com/grafana/alloy). As such, `feat` PRs are unlikely to be accepted unless a case can be made for the feature actually being a bug fix to existing behavior.
- [ ] Changes that require user attention or interaction to upgrade are documented in `docs/sources/setup/upgrade/_index.md`
- [ ] If the change is deprecating or removing a configuration option, update the `deprecated-config.yaml` and `deleted-config.yaml` files respectively in the `tools/deprecated-config-checker` directory. [Example PR](https://github.com/grafana/loki/pull/10840/commits/0d4416a4b03739583349934b96f272fb4f685d15)